Update README to note the drawbacks of using os_log from swift-log using the linked backend
While in principle, swift-log can be used to interface with os_log to log on Apple devices such as iOS platforms, doing so comes at a cost (as of now). The README file links to an implementation of a backend that uses os_log. However, the backend always uses `%@` as the format string and eagerly converts all string interpolations to Strings.  This has two drawbacks: 1. the static components of the string interpolation would be eagerly copied by the unified logging system, which will result in loss of performance. 2. It also changes the default privacy scheme used by os_log and make all message private, which runs the risk of no message getting  logged. Clarify this in the README to capture this trade off. Also, link to the on-going effort on improving os_log Swift APIs and aligning it with swift-log.